requirements.txt
django==3.1.12
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.0.0
drf-yasg==1.21.8
django-environ==0.11.2
djongo==1.3.6
pymongo==3.11.4
